feat(next-app-router): listen to external URL changes
