+++ packages/react-instantsearch-nextjs/src/InstantSearchNext.tsx
@@ -1,7 +1,4 @@
-import historyRouter from 'instantsearch.js/es/lib/routers/history';
 import { safelyRunOnBrowser } from 'instantsearch.js/es/lib/utils';
-import { headers } from 'next/headers';
-import { usePathname, useSearchParams, useRouter } from 'next/navigation';
 import React, { useEffect, useRef } from 'react';
 import {
   InstantSearch,
@@ -11,6 +8,7 @@ import {
 
 import { InitializePromise } from './InitializePromise';
 import { TriggerSearch } from './TriggerSearch';
+import { useInstantSearchRouting } from './useInstantSearchRouting';
 import { warn } from './warn';
 
 import type { InitialResults, StateMapping, UiState } from 'instantsearch.js';
@@ -47,10 +45,6 @@ export function InstantSearchNext<
   routing: passedRouting,
   ...instantSearchProps
 }: InstantSearchNextProps<TUiState, TRouteState>) {
-  const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const router = useRouter();
-
   const isMounting = useRef(true);
   useEffect(() => {
     isMounting.current = false;
@@ -60,53 +54,13 @@ export function InstantSearchNext<
     };
   }, []);
 
+  const routing = useInstantSearchRouting(passedRouting, isMounting);
+
   const promiseRef = useRef<PromiseWithState<void> | null>(null);
 
   const initialResults = safelyRunOnBrowser(
     () => window[InstantSearchInitialResults]
   );
-
-  const routing: InstantSearchProps<TUiState, TRouteState>['routing'] =
-    passedRouting && {};
-  if (routing) {
-    let browserHistoryOptions: Partial<BrowserHistoryArgs<TRouteState>> = {};
-
-    browserHistoryOptions.getLocation = () => {
-      if (typeof window === 'undefined') {
-        const url = `${
-          headers().get('x-forwarded-proto') || 'http'
-        }://${headers().get('host')}${pathname}?${searchParams}`;
-        return new URL(url) as unknown as Location;
-      }
-
-      if (isMounting.current) {
-        return new URL(
-          `${window.location.protocol}//${window.location.host}${pathname}?${searchParams}`
-        ) as unknown as Location;
-      }
-
-      return window.location;
-    };
-    browserHistoryOptions.push = function push(
-      this: ReturnType<typeof historyRouter>,
-      url
-    ) {
-      // This is to skip the push with empty routeState on dispose as it would clear params set on a <Link>
-      if (this.isDisposed) {
-        return;
-      }
-      router.push(url, { scroll: false });
-    };
-
-    if (typeof passedRouting === 'object') {
-      browserHistoryOptions = {
-        ...browserHistoryOptions,
-        ...passedRouting.router,
-      };
-      routing.stateMapping = passedRouting.stateMapping;
-    }
-    routing.router = historyRouter(browserHistoryOptions);
-  }
 
   warn(
     false,
